requirements.txt
aiohttp
aiosqlite
discord.py
python-dotenv
python-dotenv
python-a2s
asyncpraw
praw
beautifulsoup4